fix(aia): refine IMSIC CSR interfaces
Expose the standard eidelivery values and use usize for the eip and eie
selector bases. Document the extension, privilege, and guest-file
preconditions for unsafe M-, S-, and VS-mode indirect accesses.

Keep the in-tree Prototyper buildable by restoring only AddressLayout,
machine_interrupt_file_address, and encode_le from the removed IMSIC
helpers. Leave unrelated legacy utilities removed and record the change
in the changelog.

+++ library/riscv-aia/src/csrind/eie.rs
@@ -1,7 +1,7 @@
 //! External interrupt-enable registers.
 
-/// Indirect CSR identifier for `eie[n]`. registers (eie0 to eie63).
-pub const EIE_BASE: u32 = 0xC0;
+/// Indirect CSR base identifier for `eie[n]` registers (eie0 to eie63).
+pub const EIE_BASE: usize = 0xC0;
 
 /// External interrupt-enable bits register (`eie[n]`).
 ///
